Module scaleless_music::sound

Basic sound synthesizer routines.

Reexports

pub use self::amplitude::*;

pub use self::amplitude_overtones::*;

pub use self::backend_portaudio::*;

pub use self::errors::*;

pub use self::frequency::*;

pub use self::interval::*;

pub use self::note::*;

pub use self::progress::*;

pub use self::rhythm::*;

pub use self::timing::*;

pub use self::wave::*;

Modules

amplitude
Fuctions which provide amplitude changes.

amplitude_overtones
Fuctions which provide amplitude changes for overtones also.

backend_portaudio
PortAudio backend for sound playback.

errors
Error messages.

frequency
Fuctions which provide frequency changes.

interval
Frequency interval.

note
Musical note structures.

progress
Time and tempo based progress measurement.

rhythm
Rhythm section.

timing
Timing for the duration of sound components.

wave
Fuctions which provide complete waveforms.

Constants

BUFFER_SIZE_DEFAULT
Sample count for calculations. It affects both latency and computation performance. Latency perception for musical instruments: over ~12ms is already disturbing for some players.

PI2
= π x 2

SAMPLE_RATE_DEFAULT
= 192 kHz, as humans can discern time differences of 5 microseconds. Humans can hear < 1° difference in the location of the sound source, when it is in front of them. See also: interaural time difference

TONE_FREQUENCY_MAX
= 28 kHz, the highest hearable (feelable) frequency. Overtones above this frequency are filtered out from calculations. See: hearing range

TONE_FREQUENCY_MIN
= 3 Hz, the lowest feelable frequency. Tones below it will not be calculated. The hearable lowest is 12 Hz. See: hearing range

Traits

MusicStructure
A structure of music.

SoundGenerator
Sound sample generator for output (playback). It can also take real-time input (commands), thus musical instruments can be realized with it.

SoundStructure
A sound component. Can be a simple wave or a complex structure of waves.

Functions

get_sample_time
Calculates the period of one sample for the given sample rate. Sample rates below 1.0 are considered to be invalid.

Type Definitions

SampleCalc
Precision of calculations. Changing it to f64 can slow down some calculations 4 times.

SampleOutput
Precision of the finally produced samples.