Module scaleless_music::sound::amplitude

Fuctions which provide amplitude changes.

Structs

AmplitudeCombination
Combination of several amplitude functions.

AmplitudeConst
Constant amplitude.

AmplitudeDecayExp
Amplitude is decaying exponentially. The decay rate only depends on time, even when the duration is tempo dependent. Exponential decay

AmplitudeEqualLoudness
Equal-loudness contour data used is described by the ISO 226:2003 standard see also: https://plot.ly/~mrlyule/16/equal-loudness-contours-iso-226-2003/

AmplitudeSequence
Sequence of several amplitude functions.

FadeLinear
Linearly changing amplitude.

Tremolo
Tremolo, as sine variation of the amplitude.

Traits

AmplitudeJoinable
The AmplitudeJoinable trait is used to specify the ability of joining amplitude structures together, forming a sequence of them.

AmplitudeProvider
Provides time dependent amlitude changes.

Functions

is_valid_amplitude
Checks if the given value is in the valid amplitude range.