Trait scaleless_music::sound::amplitude::AmplitudeJoinable

pub trait AmplitudeJoinable: AmplitudeProvider + HasTimer {
    fn set_amplitude_start(&self, amplitude: SampleCalc) -> SoundResult<()>;
    fn get_amplitude(&self) -> SampleCalc;
    fn get_max(&self) -> SampleCalc;
}

The AmplitudeJoinable trait is used to specify the ability of joining amplitude structures together, forming a sequence of them.

Required Methods

fn set_amplitude_start(&self, amplitude: SampleCalc) -> SoundResult<()>

Sets the initial amplitude, and resets time.

fn get_amplitude(&self) -> SampleCalc

Provides the actual amplitude value.

fn get_max(&self) -> SampleCalc

Provides the maximal possible future amplitude (for normalization).