Module scaleless_music::sound::amplitude_overtones

Fuctions which provide amplitude changes for overtones also.

Structs

AmplitudeConstOvertones
Amplitude is not changing by time, this function gives the overtone amplitudes too.

AmplitudeDecayExpOvertones
Amplitude is decaying exponentially, also for overtones Exponential decay index: 0 = fundamental tone, 1.. = overtones.

AmplitudeOvertonesSequence
A sequence of amplitude functions with overtones.

Traits

AmplitudeOvertonesJoinable
The AmplitudeOvertonesJoinable trait is used to specify the ability of joining amplitude structures (with overtones) together, forming a sequence of them.

AmplitudeOvertonesProvider
Provides time dependent amlitude changes both for the fundamental tone and for overtones.