Struct scaleless_music::sound::backend_portaudio::SoundInterface

pub struct SoundInterface<T: 'static> { /* fields omitted */ }

This is a wrapper around the sound output backend

Methods

impl<T> SoundInterface<T>

fn new(sample_rate: u32, buffer_size: usize, channel_count: u16, generator: Box<SoundGenerator<Command=T>>) -> BackendResult<SoundInterface<T>>

Creates a new backend for sound playback. At the moment all channels output the same sound.

fn start(&mut self) -> BackendResult<()>

Starts the sound output stream.

fn send_command(&mut self, command: T) -> BackendResult<()>

Sends a command to the sound generator.

fn get_sample_rate(&self) -> u32

Returns the sample rate of the sond output

fn get_channel_count(&self) -> u16

Returns the channel count of the sond output