Enum scaleless_music::sound::errors::Error

pub enum Error {
    Backend(BackendError),
    SampleRateInvalid,
    BufferSize,
    OvertoneCountInvalid,
    NumeratorInvalid,
    DenominatorInvalid,
    FrequencyTooLow,
    FrequencyTooHigh,
    FrequencyInvalid,
    FrequencySource,
    RateInvalid,
    AmplitudeInvalid,
    AmplitudeTimeInvalid,
    AmplitudeRateInvalid,
    PeriodInvalid,
    DurationInvalid,
    ChannelInvalid,
    TempoInvalid,
    TimingInvalid,
    ProgressInvalid,
    ProgressCompleted,
    ItemsCompleted(usize),
    SequenceEmpty,
    ItemInvalid,
    Overflow,
}

Error types of the sound module.

Variants
Backend(BackendError)

Sound output backend error.

SampleRateInvalid

Invalid sample rate.

BufferSize

Invalid buffer size for the given sample count.

OvertoneCountInvalid

Overtone count does not match the reserved array size.

NumeratorInvalid

Numerator cannot be 0, because frequencies can not be 0.

DenominatorInvalid

Denominator cannot be 0 (division by zero error).

FrequencyTooLow

The frequency is below the hearing range.

FrequencyTooHigh

The frequency exceeds the hearing range.

FrequencyInvalid

Frequency can not be zero or negative.

FrequencySource

This frequency function is a source, it can not use an input frequency buffer.

RateInvalid

A rate must be positive.

AmplitudeInvalid

Amplitude cannot be negative.

AmplitudeTimeInvalid

Amplitude change time is not positive.

AmplitudeRateInvalid

Amplitude change rate is out of the range allowed for the given function.

PeriodInvalid

A time period must be positive.

DurationInvalid

A time duration must be positive.

ChannelInvalid

Channel of the given number does not exist.

TempoInvalid

Beats per minute must be positive.

TimingInvalid

Timing option does not match the method.

ProgressInvalid

The selected progress option is invalid for this case.

ProgressCompleted

Progress is finished.

ItemsCompleted(usize)

The number of items completed in an unfinished buffer operation.

SequenceEmpty

The Sequence has no items.

ItemInvalid

Item at the given index does not exist.

Overflow

Overflow occured during calculations.