Trait scaleless_music::sound::progress::Progress

pub trait Progress: HasTimer {
    fn set_period_unit(&self, period_unit: SampleCalc);
    fn set_phase_init(&self, phase: SampleCalc);
    fn simplify(&self);
    fn next_by_time(&self) -> SoundResult<SampleCalc>;
    fn next_by_tempo(&self, tempo: SampleCalc) -> SoundResult<SampleCalc>;
    fn get_phase(&self) -> SampleCalc;
}

Common methods of the Progress types.

Required Methods

fn set_period_unit(&self, period_unit: SampleCalc)

Sets a new period unit, and restarts the progress. Period unit is the amount of phase change during one period.

fn set_phase_init(&self, phase: SampleCalc)

Sets a new initial phase value, and restarts the progress.

fn simplify(&self)

Simplifies the phase to achieve higher accuracy. It is only used for periodic functions.

fn next_by_time(&self) -> SoundResult<SampleCalc>

Provides the next phase value, or Error::ProgressCompleted if the progress is finished.

fn next_by_tempo(&self, tempo: SampleCalc) -> SoundResult<SampleCalc>

Provides the next phase value depending on the actual tempo, or Error::ProgressCompleted if progress is finished. Tempo value is given in beats per second.

fn get_phase(&self) -> SampleCalc

Returns the actual phase value.