Enum scaleless_music::sound::timing::TimingOption

pub enum TimingOption {
    None,
    TimeConst(SampleCalc),
    TimeRatio {
        ratio: SampleCalc,
        duration: SampleCalc,
    },
    TempoConst(NoteValue),
    TempoRatio {
        ratio: NoteValue,
        duration: NoteValue,
    },
}

Optional duration type, for timings in sequences.

Variants
None

Timing is turned off (= unlimited duration)

TimeConst(SampleCalc)

Constant amount of time independent of rhythm. For a sequence item, it is independent of the sequence's whole duration.

TimeRatio

Proportion relative to a sequence's whole duration (in time). The sequence's timing must be time based!

Fields

ratio: SampleCalc

Proportion.

duration: SampleCalc

The calculated duration value (from the ratio).

TempoConst(NoteValue)

Duration relative to the beat duration. For a sequence item, it is independent of the sequence's whole duration.

TempoRatio

Proportion relative to a sequence's whole duration. The sequence's timing must be tempo based!

Fields

ratio: NoteValue

Proportion.

duration: NoteValue

The calculated duration in beats.