Trait scaleless_music::sound::timing::HasTimer

pub trait HasTimer {
    fn set_timing(&self, timing: TimingOption) -> SoundResult<()>;
    fn get_timing(&self) -> TimingOption;
    fn restart(&self);
    fn apply_parent_timing(&self, parent_timing: TimingOption) -> SoundResult<()>;
}

It provides the timing functionality required for making sequences.

Required Methods

fn set_timing(&self, timing: TimingOption) -> SoundResult<()>

Sets the timing (duration) of the sequence item, and restarts the internal timer.

fn get_timing(&self) -> TimingOption

Provides the TimingOption.

fn restart(&self)

Restarts the internal timer.

fn apply_parent_timing(&self, parent_timing: TimingOption) -> SoundResult<()>

Applies the parent's timing to calculate it's own relative timing. It is used for sequence items.